Better handling of graph names and iteration over graphs
